issue #223: update tests
